Implement scheduling algorithm for testing.
Add a new api route that allows inputs from the form to calculate a
schedule. It is in a primitive stage and therefore not production ready.

In addition to the algorithm implementation, a linter is also performed,
most style issues have been fixed automatically, but more manual changes
are to be expected towards the end of the project.